+++ src/UIComponents/atom/Checkbox/Checkbox.js
@@ -3,13 +3,13 @@ import styled, { css } from "styled-components";
 import classNames from "classnames";
 import PropTypes from "prop-types";
 
-import { baseCSS, dotCSS, getStyles } from "../../styled";
+import { baseCSS, flexCSS, dotCSS, getStyles } from "../../styled";
 
 const StyledCheckbox = styled.div`
   width: max-content;
-  display: inline-flex;
   position: relative;
   ${baseCSS};
+  ${flexCSS};
   ${(props) => getStyles({ ...props, componentName: "CHECKBOX" })};
   &.checkbox-selected {
     &:after {
@@ -37,7 +37,7 @@ const Checkbox = ({
   const classes = classNames({
     checkbox: !skipDefaultClass,
     [className]: !!className,
-    ["checkbox-selected"]: value,
+    "checkbox-selected": value,
     disabled,
   });