+++ src/UIComponents/atom/Input/Input.js
@@ -10,9 +10,6 @@ import { baseCSS } from "../../styled";
 const StyledInput = styled.input`
   background: ${colors.white};
   border: 1px solid ${constants.BG};
-  display: inline-flex;
-  flex-shrink: 0;
-  flex-grow: 0;
   ${baseCSS};
   margin: 0;
   &:active:not(.disabled),